Move getRouteForVaa from util to Operator
