+++ wormhole-connect/src/utils/routes/index.ts
@@ -6,4 +6,3 @@ export * from './bridge';
 export * from './relay';
 export * from './hashflow';
 export * from './types';
-export * from './utils';
